Align CLI gates with finalized carrier contracts

+++ test/cli-build/expressions-control-flow-functions-this-async.test.mjs
@@ -270,7 +270,7 @@ test("CLI emits omitted function and method return types from TSTS inferred sign
   assert.equal(build.status, 0, build.stderr);
 
   const generatedSource = await readFile(resolve(projectDirectory, "out/csharp/src/Index.cs"), "utf8");
-  assert.match(generatedSource, /public static double inferred\(\)/);
+  assert.match(generatedSource, /public static int inferred\(\)/);
   assert.match(generatedSource, /public static void sideEffect\(double value\)/);
   assert.match(generatedSource, /public static double\[\] inferredArray\(\)/);
   assert.match(generatedSource, /return new double\[\] \{ 1, 2 \};/);
